Simplified module pysmt.parsing into a unique file
The pysmt.parsing module was originally divided in two parts, with the
idea that the pratt parser could be used independently of pySMT. This
however turned out not to be possible.
